+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0 -- UNRELEASED
 
 ### FEATURES
-- ...
+- first version of Finance complete
 
 ### ENHANCEMENTS
 - ...